\name{tsdata}
\alias{tsdata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Time Series Data for Testing
}
\description{
A simulated time series data model generates trajectories in a two-dimensional space. Generated trajectories are composed of 20-observation cycles, each cycle with four quarter-ellipse segments of five observations.
}
\usage{
data(tsdata)
}
%- maybe also 'usage' for other objects documented here.

\value{
Provides user data to run as an example
}
\references{
Caudle, KA, Fleming, PS, Frey, MR and Brubaker, N. "Next Generation of Flow Field Forecasting", Proceedings of the Joint Statistical Meetings of the American Statistical Association, Seattle, WA., 8 August-13 August (2015).

Frey, Michael R., and Kyle A. Caudle. "Flow field forecasting for univariate time series." Statistical Analysis and Data Mining (2013).
}
\author{
Patrick Fleming
}

\examples{
data(tsdata) # Load time series data int R
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
